debug Model

+++ TrainModel.py
@@ -24,41 +24,42 @@ dir_name = r'./CheckPoin/'
 
 # Store layers weight & bias
 image_weights = {
-    'wc1': tf.Variable(tf.random_normal([15, 15, 3, 64], dtype = tf.float64)),
-    'wc2': tf.Variable(tf.random_normal([5, 5, 64, 128], dtype = tf.float64)),
-    'wc3': tf.Variable(tf.random_normal([3, 3, 128, 256], dtype = tf.float64)),
-    'wc4': tf.Variable(tf.random_normal([3, 3, 256, 256], dtype = tf.float64)),
-    'wc5': tf.Variable(tf.random_normal([3, 3, 256, 256], dtype = tf.float64)),
-    'wd1': tf.Variable(tf.random_normal([8*8*256, 512], dtype = tf.float64)), 
-    'wd2': tf.Variable(tf.random_normal([512, 256], dtype = tf.float64)),
+    'wc1': tf.Variable(tf.random_normal([15, 15, 3, 64])),
+    'wc2': tf.Variable(tf.random_normal([5, 5, 64, 128])),
+    'wc3': tf.Variable(tf.random_normal([3, 3, 128, 256])),
+    'wc4': tf.Variable(tf.random_normal([3, 3, 256, 256])),
+    'wc5': tf.Variable(tf.random_normal([3, 3, 256, 256])),
+    'wd1': tf.Variable(tf.random_normal([8*8*256, 512])), 
+    'wd2': tf.Variable(tf.random_normal([512, 256])),
 }
+
 image_biases = {
-    'bc1': tf.Variable(tf.random_normal([64], dtype = tf.float64)),
-    'bc2': tf.Variable(tf.random_normal([128], dtype = tf.float64)),
-    'bc3': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
-    'bc4': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
-    'bc5': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
-    'bd1': tf.Variable(tf.random_normal([512], dtype = tf.float64)),
-    'bd2': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
+    'bc1': tf.Variable(tf.random_normal([64])),
+    'bc2': tf.Variable(tf.random_normal([128])),
+    'bc3': tf.Variable(tf.random_normal([256])),
+    'bc4': tf.Variable(tf.random_normal([256])),
+    'bc5': tf.Variable(tf.random_normal([256])),
+    'bd1': tf.Variable(tf.random_normal([512])),
+    'bd2': tf.Variable(tf.random_normal([256]),
 }
 
 sketch_weights = {
-    'wc1': tf.Variable(tf.random_normal([15, 15, 3, 64], dtype = tf.float64)),
-    'wc2': tf.Variable(tf.random_normal([5, 5, 64, 128], dtype = tf.float64)),
-    'wc3': tf.Variable(tf.random_normal([3, 3, 128, 256], dtype = tf.float64)),
-    'wc4': tf.Variable(tf.random_normal([3, 3, 256, 256], dtype = tf.float64)),
-    'wc5': tf.Variable(tf.random_normal([3, 3, 256, 256], dtype = tf.float64)),
-    'wd1': tf.Variable(tf.random_normal([8*8*256, 512], dtype = tf.float64)), 
-    'wd2': tf.Variable(tf.random_normal([512, 256], dtype = tf.float64)),
+    'wc1': tf.Variable(tf.random_normal([15, 15, 3, 64])),
+    'wc2': tf.Variable(tf.random_normal([5, 5, 64, 128])),
+    'wc3': tf.Variable(tf.random_normal([3, 3, 128, 256])),
+    'wc4': tf.Variable(tf.random_normal([3, 3, 256, 256])),
+    'wc5': tf.Variable(tf.random_normal([3, 3, 256, 256])),
+    'wd1': tf.Variable(tf.random_normal([8*8*256, 512])), 
+    'wd2': tf.Variable(tf.random_normal([512, 256])),
 }
 sketch_biases = {
-    'bc1': tf.Variable(tf.random_normal([64], dtype = tf.float64)),
-    'bc2': tf.Variable(tf.random_normal([128], dtype = tf.float64)),
-    'bc3': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
-    'bc4': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
-    'bc5': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
-    'bd1': tf.Variable(tf.random_normal([512], dtype = tf.float64)),
-    'bd2': tf.Variable(tf.random_normal([256], dtype = tf.float64)),
+    'bc1': tf.Variable(tf.random_normal([64])),
+    'bc2': tf.Variable(tf.random_normal([128])),
+    'bc3': tf.Variable(tf.random_normal([256])),
+    'bc4': tf.Variable(tf.random_normal([256])),
+    'bc5': tf.Variable(tf.random_normal([256])),
+    'bd1': tf.Variable(tf.random_normal([512])),
+    'bd2': tf.Variable(tf.random_normal([256]),
 }
 
 
@@ -67,9 +68,9 @@ def EuclideanDist(a, b):
 
 def run_training():
     
-    sketchs_placeholder = tf.placeholder(tf.float64)
-    images_neg_placeholder = tf.placeholder(tf.float64)
-    images_pos_placeholder = tf.placeholder(tf.float64)
+    sketchs_placeholder = tf.placeholder(tf.float32)
+    images_neg_placeholder = tf.placeholder(tf.float32)
+    images_pos_placeholder = tf.placeholder(tf.float32)
     keep_prob = tf.placeholder(tf.float64)
 
     # Three Branch Net
@@ -83,11 +84,11 @@ def run_training():
     # Euclidean Distance
     dist_pos = EuclideanDist(sketch_dense, image_pos_dense)
     dist_neg = EuclideanDist(sketch_dense, image_neg_dense)
-    margins = tf.constant(margin, dtype = tf.float64, shape = [batch_size])
+    margins = tf.constant(margin, dtype = tf.float32, shape = [batch_size])
     print(dist_pos, dist_neg, margins)
 
     with tf.name_scope('Loss') as scope:
-        zeros = tf.constant(0.0, dtype = tf.float64, shape = [batch_size])
+        zeros = tf.constant(0.0, dtype = tf.float32, shape = [batch_size])
         cost = tf.reduce_sum( tf.maximum(zeros, margins + dist_pos - dist_neg) )
         tf.summary.scalar("loss", cost)
